src/cloud-services/shared-services/health-reliability-monitoring/requirements.txt
# Health & Reliability Monitoring Dependencies

# Web framework
fastapi>=0.109.0
uvicorn[standard]>=0.25.0
pydantic>=2.6.1

# Database & migrations
sqlalchemy>=2.0.25
alembic>=1.13.1
psycopg2-binary>=2.9.9

# Async tasks & messaging
redis>=5.0.1
aiokafka>=0.10.0

# Telemetry & metrics
opentelemetry-sdk>=1.23.0
opentelemetry-exporter-otlp>=1.23.0
prometheus-client>=0.19.0

# Serialization & perf
orjson>=3.9.10
python-dateutil>=2.8.2

# Security
PyJWT>=2.8.0
cryptography>=42.0.0

# HTTP client
httpx>=0.26.0

# Testing
pytest>=7.4.4
pytest-asyncio>=0.23.4
pytest-cov>=4.1.0
pytest-mock>=3.12.0
k6>=0.0.0  # placeholder for load scripts reference

